+++ cspell.json
@@ -42,6 +42,7 @@
     "dspc",
     "embedder",
     "embedders",
+    "envfiles",
     "estree",
     "euclidian",
     "expfmt",
